Open the Tomb's north door onto The Pit
The Tomb's third door has been drawn and sealed since the room was built.
The Pit is what goes behind it: Chapter 3's sixth room and the first in the
chapter with two walkable heights that are not a maze. A shaft 4800 by 3900,
open top to bottom, with a gallery at y=-1000 running in a U round three
sides and the pit floor 1000 units below at y=0. Only the gallery is
reachable — the player walks in through the south door, walks the U, looks
down into the shaft and walks back out. The way down is a later job, and
both the header and the .c are written so adding it is a floor zone and a
wall rather than a re-think. The two levels have disjoint XZ footprints, so
the door trigger stays a plain Manhattan test with no storey check.

The north door needed a third sign pairing: it is an XY-plane door at fixed
Z, so the mirror rule and the reading axis both differ from the Tomb's two
YZ doors. tomb.h now spells all three out at their call sites.

The Pit is the thirty-third room, which is exactly the overflow tomb.h's
predecessor predicted: WORLD_NUM_ROOMS 32 filled the uint32_t `visited`
mask exactly, the static assert fired, and `visited` and `cribs_solved`
are both uint64_t now. The ceiling is 64. No SAVE_VERSION bump was needed.

New rusty texture in TEXBANK_CATACOMBS, with the VRAM maps, texture-bank
table and heap budget updated to match.

+++ src/crib.h
@@ -409,7 +409,9 @@ void cribs_collide(int32_t *px, int32_t py, int32_t *pz, int32_t radius);
  * live every frame rather than caching the answer in the instance, so the order
  * of room init, world_enter() and savegame_apply_pending() cannot matter.
  * See the long note at the top of this file for why this is not a GameFlag. */
-uint32_t crib_solved_mask(void);
-void     crib_solved_mask_set(uint32_t mask);   /* load; replaces the whole set */
+/* uint64_t since The Pit took room slot 32 and widened WorldDelta.visited and
+ * cribs_solved beside it (src/world.h). One bit per room, ceiling 64. */
+uint64_t crib_solved_mask(void);
+void     crib_solved_mask_set(uint64_t mask);   /* load; replaces the whole set */
 int      crib_room_solved(GameState area);
 #endif
